+++ packages/base/Note/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @toptal/picasso-note
 
+## 4.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`c40f4aa`](https://github.com/toptal/picasso/commit/c40f4aa6b465a22b54a316c1088b59cd63724b9d)]:
+  - @toptal/picasso-typography@5.1.0
+
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`440f217`](https://github.com/toptal/picasso/commit/440f217c1748d09beeca90e5277d2137d4251897)]:
+  - @toptal/picasso-container@3.1.5
+  - @toptal/picasso-typography@5.0.1
+  - @toptal/picasso-tailwind-merge@2.0.5
+
 ## 4.0.7
 
 ### Patch Changes