+++ packages/base/Tabs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @toptal/picasso-tabs
 
+## 7.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`c40f4aa`](https://github.com/toptal/picasso/commit/c40f4aa6b465a22b54a316c1088b59cd63724b9d)]:
+  - @toptal/picasso-typography@5.1.0
+  - @toptal/picasso-typography-overflow@4.0.8
+  - @toptal/picasso-user-badge@5.1.24
+
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`440f217`](https://github.com/toptal/picasso/commit/440f217c1748d09beeca90e5277d2137d4251897)]:
+  - @toptal/picasso-shared@16.0.0
+  - @toptal/picasso-container@3.1.5
+  - @toptal/picasso-icons@1.15.3
+  - @toptal/picasso-typography@5.0.1
+  - @toptal/picasso-utils@4.0.1
+  - @toptal/picasso-user-badge@5.1.23
+  - @toptal/picasso-typography-overflow@4.0.7
+
 ## 7.0.1
 
 ### Patch Changes