+++ packages/base/Utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-utils
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`440f217`](https://github.com/toptal/picasso/commit/440f217c1748d09beeca90e5277d2137d4251897)]:
+  - @toptal/picasso-shared@16.0.0
+
 ## 4.0.0
 
 ### Major Changes