+++ packages/picasso-tailwind-merge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-tailwind-merge
 
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-utils@4.0.1
+
 ## 2.0.4
 
 ### Patch Changes